grub: put kernel symlinks in /EFI/ubuntu instead of the root dir
The root dir is not mounted at runtime on top of /boot/grub, so we aren't able
to see the kernel.efi symlink when it is placed in the root directory of the
ubuntu-boot partition. Instead put the symlink in /EFI/ubuntu/kernel.efi, so
that at runtime this shows up as /boot/grub/kernel.efi and is more easily
manipulated.

Also update tests with these changes.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>