render: fix overwriting of double with runes
While working on the algorithm yesterday I noticed that some situations
when double or single width rune overwrites a double or single width
rune or the empty space after each double with rune would not be handled
correctly.

I fixed the compositing algorithm and added tests that illustrate this
operation and how it, now, correctly behaves in such situations.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>