grub.go: simplify comments, use Lstat to check if symlink exists
We really only care if the symlink exists, not if the symlink points to a valid
destination or not. If the symlink doesn't exist then we are probably trying to
disable the wrong thing, but if the symlink is dangling, then we might be
cleaning up after an issue with a try-kernel.

Also os.Remove is simpler to use and understand than osutil.UnlinkMany, so use
that instead.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>