interfaces/apparmor: don't omit deny rules in devmode
Also split out deny rules from the default policy template so that those rules
also aren't inserted into devmode policies.

Note: we specifically only want to omit these deny rules when we are in devmode,
so if we are in classic and in jailmode, then we still include these rules.

Finally, don't omit the deny ptrace rules if the snap is in devmode, because a
good number of complicated snaps, such as anything container related or even
standalone applications like bandwhich, require accesses which trigger ptrace
capabilities, so not including this snippet when run in devmode will simplify
the process of strictly confining these applications.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>