tests: switch mount-ns test to differential data set
This change comes from a realization that most of the maintenance cost
comes from the fact that any change in the middle of a mount table
shifts the re-numbered values of various identifiers of all the elements
following the change. To counter this, instead of storing absolute
renumbered values, store differential, renumbered values.

This is achieved in several steps:
  - Introduce a signed integer type, that acts like a normal integer but
    renders positive values with a plus sign. This is bulk of the
    rendering change that was required.
  - Alter Device type to be a regular object and not a subclass of int.
    This is required because it is impossible to represent -1:-1 or
    other similar values the "regular" way without gross hacks.
  - Add routines for diffing MountInfoEntry and OptionalFields
  - Add --differential command line option that computes a differential
    mount entry, relative to previous entry even if that entry is not
    displayed.

All the mount tables are, by necessity, re-generated in their
differential form. The ubuntu classic 16.04 tables were more stale
as they were disabled before.

This patch is improved to fix a bug in diffing optional fields. Thanks
to Ian for spotting!

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>