gadget: skip update when raw structure content is unchanged
Even though the structure edition is bumped, the actual content of raw structure
may be unchanged. This can only be determined once we run the `Backup()` phase
of the update sequence. Should such scenario be identified, return an error
indicating that there is no update. The callers up the stack are already
prepared to handle such scenarios and trigger no reboot.

The mechanism is introduced for 'raw' (or non-filesystem) structures first,
since their analysis is much simpler.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>