cmd/snap-bootstrap/initramfs_mounts: mount kernel snap from the bootloader
This is necessary because the initramfs will mount the kernel snap to load
kernel modules that are necessary to proceed in booting, for example network
related modules needed to access the device.

Currently this follows the same logic that snapd follows in determining which
kernel we booted, which is not robust against tampering because ubuntu-boot is
not a trusted partition, however we may not be able to protect against this
attack without a hardware root of trust to measure the boot chain. If we enforce
secure boot, we can at least block the attack of inserting arbitrary kernel.efi
images onto ubuntu-boot, but this code would still suffer from incorrectly
choosing which kernel.efi we booted if an attacker wanted to tamper with the
kernel.efi on the ubuntu-boot partition.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>