grub: support atomically renaming kernel symlinks
Previously, makeKernelEfiSymlink would fail if the symlink already existed. This
situation happens when we have a kernel snap refresh, where the kernel.efi
symlink already exists and we want to move it after a successful refresh. We can
fix this by using the AtomicSymlink helper, which atomically creates the
symlink, even if the symlink already exists.

Also add a unit test for this situation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>