config: add system.certs.[a-zA-Z0-9] support
This option allows adding pem encoded ssl certificates
to the snapd config that will then be used by snapd.

Certs are added via:
```
$ sudo snap set system certs.extra1="$(cat mykey.pem)"
```
and removed via:
```
$ sudo snap unset system 'certs.extra1'
```

The certs are written to /var/lib/snapd/ssl/certs
and picked up by the httputil.NewHTTPClient(). This
means that they will be picked up by snap-repair as
well automatically.

Snapd on classic will also honor the ssl config but
it's less useful here because there are distro mechanisms
already to add certs.

One downside of this PR as is is that the certs end up in the
state. To fix this we will need some more work in the config
system AFAICT.