bootstate20.go: refactor commit() to error with both snapToTry and snapTried set
This ensures that if we do somehow get into a weird situation where both
setNext() and markSuccessful() are called on the same bootStateUpdate20, then we
will fail here with an error message, rather than errantly enable a kernel.

Additionally, we will always disable the try kernel as long as kernel_status in
the bootloader isn't "". This should be logically the same as before, but is
perhaps more explicit now.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>