daemon, store: better expose single action errors
When a snap action (that is, a POST to the store's refresh endpoint)
results in a single error for a single operation on a single snap, the
current code raises it as an InternalError.

This means a download of an unknown snap becomes an internal error; a
download of a snap from the wrong channel becomes an internal error,
etc.  This relatively minor issue becomes user-visible wackiness with
the switch to download being via snapd (and using the mentioned
endpoint).

This change fixes that, such that _if_ it's a single operation on a
single snap returning a single error, then the error raised is the
internal one.

More work could be done to coalesce things such that if it's the same
error for multiple snaps it's handled differently; this would simplify
the "no refreshes found" logging and reporting, I suspect (but haven't
checked).