snap-bootstrap/cmd_initramfs_mounts.go: handle base snap updates in the modeenv
Now, snap-bootstrap will read the "base_status" and "try_base" from the modeenv,
and if those are set, it will inform the initramfs to mount the base snap from
the "try_base" instead of the "base".

This also means we need to write the modeenv again when we run in run-mode,
because the initramfs needs to change base_status=try -> base_status=trying.

Still some unknowns about what to do in error conditions here however, those are
left as TODO:UC20s.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>