overlord/snapstate, wrappers: manpages!
This is a rough draft of a first pass at manpages support.

This works by creating a `man` directory next to the snappy `bin`
directory (i.e. a `/snap/man` to go with `/snap/bin`, in Ubuntu); if
the latter is on your `$PATH`, then man will pick up manpages that are
in the former.

Right now it only sets things up for files that start with the snap
name. A second pass would be to add also support for aliases. Also it
only considers files, whereas often manpages will be symlinks. We
should support manpages symlinking to other things as long as they are
in the same snap, at least.

---

It works! But needs a few things before being production-ready:

* more and better unit tests
* a spread test

but I'm stopping here. Hopefully this is enough...