netutil: add default gateway monitor
This PR adds support for detecting adding/removing a default gateway.
The use-case is that we can retry registration and refreshes when a
new default gateway gets added. The assumption is that when a default
gateway gets added that this most likely means that someone configured
the device with new network configuration. This is most useful in a
first-setup situation when the device connects to e.g. WiFi but needs
to be configured first. As soon as the device is configured we should
try to get a serial in this scenario.