bootstate20: add code to clean up the kernel_status if inconsistent
If kernel_status is not empty, but we don't have a try kernel snap, we should
clean up the kernel_status variable in the bootloader, so do that here.

Also log a message when this happens in the field to aid debugging if this ever
somehow happens in the wild, it's more clear what snapd did.

Finally, add a test for this situation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>