boot: revisions() returns status string now; refactor UC20 boot var cleaning
The revisions() method now return the status string instead of whether the snap
is in a trying state, so that the caller has more information, and can do more,
for example with the markSuccessful methods for bootstate20, we can more easily
cleanup if the status is in trying mode, but there is not a try snap on disk.

Additionally, refactor genericSetNext in bootstate20 to use less return values,
as rebootRequired really is just equal to setAsTrySnap, and we can determine if
the "next" snap should be set by checking if the status returned is TryStatus.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>