cmd/snap-confine: bring /var/lib/dhcp from host, if present
The directory /var/lib/dhcp used to be available on ubuntu-core-16
systems and is exposed to snap applications by plugging the
network-control interface. We were alerted by a developer who was
porting his application from ubuntu-core-16 to ubuntu-core-18 that the
application no longer worked.

Cursory inspection revealed that the directory was only provided by the
so-called legacy mount namespace setup, where all of the root filesystem
of the host is used to construct the mount namespace of the snap. When
used with other bases the directory needs to be provided explicitly.

Fixes: https://bugs.launchpad.net/snapd/+bug/1862637
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>