bootstate20: add unit test for cleaning base_status
This was previously not working because we were setting commitBaseStatus on
bs20, when u was the thing that needed to read this.

To fix this issue, we will unconditionally set base_status to default state in
the commit() method on bootState20MarkSuccessful, and same for kernel_status.
This also has the side-effect of simplifying genericMarkSuccessful to not need
to return a shouldClean variable as well.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>