cmd/snap-bootstrap: verify kernel snap is in modeenv before mounting it
This protects us against attacks forcing an old kernel snap from having it's
modules loaded by modifying the symlink on ubuntu-boot, which is unencrypted
even in a FDE scenario on UC20.

This also protects us against attacks where the kernel.efi and try-kernel.efi
symlinks are swapped, because we know specifically which current_kernels
value is the try-kernel and the fallback kernel.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>