bootloader: make uboot a RecoveryAwareBootloader
This commit adds support for uboot to be used as a
RecoveryAwareBootloader. It works similar to what we do in grub,
i.e. it will write a recovery system specific config into the
recovery system specific dir.

Uboot has the "env import" comamnd which will be used to get
this data.