sandbox/cgroup: tweak potentially confusing walker
The walker short-circuited an error being provided by the golang
standard library _or_ a directory being visited (where error is nil)
into one condition, returning the error (either real or nil) to quickly
express the fact that such case is not interested for analysis in the
walk funciton. This was correct but possibly confusing. Split the code
into explicit chunks and tests.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>