managers_test.go: add first draft of boot robustness test
This is not yet working, but is pretty close to how this should work I
think. We set the bootloader to panic during a particular side-effect
function, and then expect that panic to happen in a particular place in
the test, then check that we were able to recover and that we cleaned up
from the failed upgrade.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>