tests/lib/prepare.sh: set StartLimitInterval=100
Setting this to 0 seems to not trigger OnFailure settings, because systemd waits
for the service to be "totally absolutely really surely failed" before running
OnFailure, and it doesn't consider the service to be in that state until systemd
has exhausted all the retries with StartLimitInterval, effectively giving the
unit a chance to fix itself by restarting before considering it failed.

It's unclear why this test ever worked before, but I think it's probable that a
systemd bug where OnFailure was run before StartLimitInterval was exhausted was
allowing this to work and now that systemd bug was fixed, hence breaking our
tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>