tests/lib/prepare.sh: use StartLimitBurst too
This is the right fix because otherwise StartLimitBurst defaults to 5, which
when combined with StartLimitInterval=100, limits us to 5 restarts every 100
seconds, which is not what we want. Instead limiting to 1000 restarts a second
should be fine.

Also add a comment about StartLimitInterval being renamed to
StartLimitIntervalSec in newer systemd versions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>