cmd/snap-failure/snapd: also rm snapd.socket if it still exists
Sometimes, we see that even after stopping the snapd.socket job, and waiting for
the snapd manually executed from the previous revision the socket file still
exists and systemctl isn't able to restart the snapd.socket job. So to get
around this and be more robust, we can try to remove that file if it exists, and
if we fail then try systemctl anyways because it might just work.

Also, don't die if we fail to reset the failed state of snapd.socket, because we
may still be able to restart the service and keep going. In general,
snap-failure should try to be as robust as it can be, because if it fails to
restore the previous revision of snapd then an UC device can be effectively
bricked.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>