snapd.failure.service.in: don't set RemainAfterExit
This specifically will cause snapd.failure to kick in if 2 broken snapd's are
installed without a reboot or explicitly stopping the snapd.failure service so
that on the second time, snapd.failure will actually be started.

Because of this, it's no longer sufficient to check to see if the snapd.failure
service is active after a broken snapd is installed, we have to look at the logs
to see that it was actually activated/started more times after the broken snapd
was installed than before.

Also since it's always basically inactive, remove the restore section since
there's no point to stopping it now as it will always be inactive essentially.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>