run-checks: SKIP_GMFMT really skips formatting checks
Previously setting SKIP_GOFMT would run the check, producing loads of
output on specific version of go, even in an otherwise pristine tree and
then IGNORE the "failure" and carry on with other checks.

This may have been useful once but it no longer is, since master has
drifted far from the "proper", fronzen-in-time version we verify
against.

We lost some time looking for the real failure (which was a unit test
error that was burried beneath the gofmt noise) and I promised I would
look at changing this today. Setting SKIP_GOFMT to a non-empty value
skips the whole formatting check.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>