travis.yml: run unit tests on arm64 as well
Use the new travis feature to run unit tests on arm64. This would
have caught the segfault in the rawstocketstopper code.

This also fixes two issues in usersession tests that fail often
in arm. One is the timeout of TestAgentTimeout that is not high
enough for slow arm systems for the "regular" case that should
not hit this timeout.

And then there is a "sleep 99999" in the mock for kdialog which
will cause travis to wait forever for this process to exit.