tests: use ipv4 in retry-network to unblock failing master
This commit unblocks master. Most recently "getenv hosts" returns
ipv6 addresses. But when those are passed to the script they are
not put inside [] so the "http://2001::01:123" style confuses
go ang things fail. I'm also working on a followup that will
use ipv6. Unfortunately this is more work as right now we get
a very different error in the network namespace and it's not
quite clear yet if the network namespace is just not configured
right or if it's a real error we should handle.