tests: add session-tool, a su / sudo replacement
During investigation of the test failure related to application tracking
for refresh-app-awareness we realized that the case of running as a
non-root user was not covered very well.

This opened the rabbit's hole of exploring what it means to run as a
user. Many of our tests use a combination of sudo and su to run a bit of
code as the user "test". This works okay for the limited application it
serves, the process runs as the non-root user but most of the
environment is still that of the original root session, that spread
creates when it uses ssh to connect to a system.

This patch introduces session-tool. This contraption is much like su or
sudo, allowing us to run a bit of code as another user. Unlike both it
does so in a proper user session.

Sessions are a fuzzy concept. There are many UNIXy things that claim to
be a session. For the purpose of this tool a session encompasses *all*
of them. As a logged in user you have access to a dbus session bus, to a
systemd --user daemon, you have XDG_RUNTIME_DIR, the session is
registered with logind, there's an appropriate cgroup hierarchy. There
may be additional user-session services running in the background. DBus
activation works and can be relied upon. When the session is terminated
all of those resources are reclaimed.

This is what session-tool provides. The details are a bit more complex
than one would hope. The tool is also rather imperfect, not offering
interactive sessions (I dropped that code for simplicity) and is
arguably not pretty. On the up side it is portable across all our test
systems, except for Ubuntu 14.04 which lacks enough systemd components
to allow us to implement a real session.

The means of crafting a session are arcane and ought to be documented
here. The main goal relies on invoking PAM and not tripping over along
the way. PAM will perform, with the aid of appropriate configuration and
modules, all of the session-y things defined above (and more, for
details do read PAM configuration of your favourite system, it is
nicely documented in manual pages). The pam-systemd module, most
notably, is responsible for handling logind interaction and giving us a
session bus. This is performed by a DBus call to logind. Interestingly
(use debug stanza in the pam_systemd.so module to find out), this will
not happen if the calling process is already a part of a logind session.
This is subsequently determined by checking UNIX peer credentials and
walking the cgroup hierarchy established by systemd. To avoid this
critical problem we interact with systemd to start a process in a new
transient service. This places is it a place that is not a session and
allows us to perform the rest. The process then execs runuser -l, which
handles PAM session handling for us. I refrained from using systemd's
native PAM integration as I found a bug present across all releases of
all OSes, where systemd cannot talk to logind due to DBus configuration
shipped by systemd itself.

The actual script uses dbus-monitor to await for the JobRemoved signal,
launches the helper service using busctl, processes the output with awk
and a few pipes to pass data around. It cannot return the precise exit
code but it can return the most relevant pass/fail information and
bridge IO streams to the calling process.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>