interfaces/udev: write udev rules but don't run if absent
Udev control socket may be absent if udev is not really running. Earlier
approach would disable the backend entirely but this would force us to
include this fact in the system-key.

This patch reverses the direction. When udev is disabled we write the
rules but not ask udev to load them. Eventually, if udev starts up, it
will load the rules anyway.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>