interfaces/seccomp: allow passing an address to setgroups
It turns out that some programs use setgroups(0, &addr) to drop groups
instead of setgroups(0, NULL). For instance, dnsmasq:
http://thekelleys.org.uk/gitweb/?p=dnsmasq.git;a=blob;f=src/dnsmasq.c;h=10f19ead62f791568532115e871a0dd7ab3527bc;hb=HEAD#l701
This might be due to compatibility reasons with BSDs. Allow this,
because if the size is 0, the second argument is not used anyway.

This is the comment from jdstrand in the review:

I read the kernel source for the setgroups() syscall for 3.2, 3.13,
4.4, 4.15, 5.0, 5.3 and 5.4 and in all cases
kernel/groups.c:groups_from_user() will use a for loop with a
terminating condition of '< count' where 'count' is the value of
'size' from userspace's invocation of setgroups() (note, the 3.2 logic
is a bit different but equivalent), so this change is safe.

That said, I would argue this is a bug in the calling program since
the setgroups man page clearly states "A process can drop all of its
supplementary groups with the call: setgroups(0, NULL);". Using an
LD_PRELOAD to address this is easy: sergiusens/snapcraft-preload#39.

As for why dnsmasq is doing this, I couldn't determine it. It was
first introduced in dnsmasq-1.7 back in July of 2002 with a changelog
entry of "1) Drop supplimentary[sic] groups" and it seems that back
then there wasn't revision control and just code drops.

In terms of pure code correctness, I don't care for this change since
dnsmasq isn't dropping in the manner that glibc
prescribes. Pragmatically, this change LGTM since the kernel is doing
the right thing (and this isn't likely to intentionally change since
it would be a pretty bad bug for the kernel to look at the group list
when size is 0 and expect it to mean anything. It could, in theory,
change to enforce that group list is NULL when size is 0, but this
PR's change would be safe in that regard since the kernel would fail
closed (but I don't think the kernel will ever change in this manner
anyway)).