wrappers: import /etc/environment in all services
We were alerted that setting a http proxy in the system is not actually
allowing snap services to see the appropriate value. On core systems
proxy settings are written to /etc/environment but systemd service units
that start all snap-installed services do not load that file. This patch
fixes that.

Fixes: https://bugs.launchpad.net/snapd/+bug/1866095
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>