uboot: use ExtractedRecoveryKernelImageBootloader
It doesn't make any sense to use RecoveryAwareBootloader, because we
can't set any recovery system on uboot without also extracting the
kernel for that recovery system. Setting the recovery system is
effectively just extracting the kernel into an arbitrarily chosen
"kernel" directory of the system recovery directory.

This also means we won't have additional uboot envs in the recovery
system.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>