cmd/snap-exec: add test case for LP:#1860369
It was noticed that snap-exec's handling of environment is incorrect.
Under certain circumstances the environment block contains more than one
entry for the given key (environment variable). In addition, though this
is not illustrated in this patch, the internal expansion code (built
into parts of snapd) uses only one of those values when expanding
environment variables ahead of forwarding their values to snap
applications.

This patch contains only the test case showing the problem.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>