snap-confine: unconditionally add /dev/net/tun to the device cgroup
When CONFIG_TUN=m, /dev/net/tun will exist but using it doesn't autoload
the tun module but also /dev/net/tun isn't udev tagged until it is
loaded. To work around this, if /dev/net/tun exists, add it
unconditionally to the cgroup and rely on AppArmor to mediate the
access.

References:
https://forum.snapcraft.io/t/raspberry-pi-3-ubuntu-core-18-network-control-interface-issue/14773/4
https://bugs.launchpad.net/snappy/+bug/1859084