interfaces: fix performance issue affecting uio interface
A system with a number of UIO slots connected to a plug would generate
multiple copies of a rule "/sys/devices/platform/**/uio/uio0/** r,",
with successive changes to the UIO device number which causes
pathologically slow compilation times in apparmor_parser.

We invoke apparmor_parser with -O no-expr-simplify which disables the
optimizer that simplifies expressions that have a common part. This
subsequently causes an explosion of the number of states and causes high
CPU and memory usage required to complete the compilation process.

We cannot enable the optimizer as that heavily affects armhf. For details
see https://gitlab.com/apparmor/apparmor/-/merge_requests/246 and
https://bugs.launchpad.net/ubuntu-rtm/+source/apparmor/+bug/1383858

As a workaround we can do two things. First of all the apparmor
permission for sysfs is relaxed so that any UIO connection allows
read-access to the meta-data of all UIO devices. This specifically
excludes the /dev/uioNNN device node so it is deemed acceptable.
Second, all of the sysfs permissions are aggregated into one apparmor
snippet that is additionally de-duplicated on the side of snapd,
preventing apparmor_parser from having to cope with duplicates.

Fixes: https://launchpad.net/bugs/1866349
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>