tests/lib/reset: workaround unicode dot in systemctl output
The newer versions of systemd seem to insist on printing out a huge colorful dot
in the output of `systemctl --failed` command:

●snap.test-snapd-service-stop-timeout.svc.service not-found failed failed snap.test-snapd-service-stop-timeout.svc.service

The dot is printed even when the output is piped to another program. Make sure
we pass --plain to avoid this. The dot is then replaced by whitespace. Adjust
the pattern to handle that.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>