apparmor: use rw for uuidd/request to default and remove from elsewhere
8537ba5b2816ac1a2f77dc03ba709947b11a2531 for PR 7779 added read-only
access to /run/uuidd/request to the snap-update-ns profile. This was
meant to be added to the default template, so in PR 8232 in commit
2049f2c8245da1d97c96fcc76bab871eb0491d23, the read-only access was added
to the default template. This access was not enough since libuuid
actually must also write the UUID_OP_* value before reading from the
socket.

Remove the unnecessary access from snap-update-ns (which should pose no
regression risk since the rule was only for read-only and thus
ineffective) and add 'w' to the access in the default template.

Also remove the now unneeded apparmor rule from openvswitch-support.

References:
https://forum.snapcraft.io/t/uuidd-apparmor-denial/16013
8537ba5b2816ac1a2f77dc03ba709947b11a2531
2049f2c8245da1d97c96fcc76bab871eb0491d23
https://launchpad.net/bugs/1867571
https://github.com/karelzak/util-linux/blob/master/libuuid/src/uuidd.h#L43-L49
https://github.com/karelzak/util-linux/blob/master/libuuid/src/gen_uuid.c#L371